Add User model
Using models allows to add some features like `plugin :update_or_create`.

Update application structure.
