+++ config/initializers/database.rb
@@ -1,0 +1,3 @@
+require 'sequel'
+
+Sequel.connect(DATABASE_URL)
