+++ main.rb
@@ -1,11 +1,4 @@
-require 'telegram/bot'
-require_relative 'db/database'
-require 'dotenv/load'
-require 'byebug'
-Dir[File.join(__dir__, 'config', 'initializers', '*.rb')].each { |file| require file }
-Dir[File.join(__dir__, 'app', 'services', '*.rb')].each { |file| require file }
-
-DB = Database.new # https://github.com/jeremyevans/sequel#the-db-convention
+require_relative 'config/application'
 
 Telegram::Bot::Client.run(TOKEN) do |bot|
   bot.listen do |message|
